advanced_web_programming/final_project/pl_api_django/requirements.txt
autopep8==1.4.4
Django==2.2.7
django-extensions==2.2.5
djangorestframework==3.10.3
pkg-resources==0.0.0
pycodestyle==2.5.0
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
